When restoring a "F" type from SaltXML make sure to use Double.
Both "Double" and "Float" are stored as "F" in SaltXML so when restoring it the one with the highest resolution should be used.

+++ salt-api/src/main/java/org/corpus_tools/salt/util/internal/persistence/SaltXML10Handler.java
@@ -340,7 +340,7 @@ public class SaltXML10Handler extends DefaultHandler2 implements SaltXML10Dictio
 				retVal = Long.parseLong(value.substring(3));
 			}
 		} else if (value.startsWith("F")) {
-			retVal = Float.parseFloat(value.substring(3));
+			retVal = Double.parseDouble(value.substring(3));
 		} else if (value.startsWith("U")) {
 			retVal = URI.createURI(value.substring(3));
 		}else if (value.startsWith("O")) {
